First version of API working with "smt configure"
Address #7

The first step is too make smt configure add a test project. It is
quite ugly the way the client does this and it needs to be fixed, but
the goal is to make the app work with the client as it stands without
changes if possible. Created a /api/v3/<project_name> view that
updates the database and returns the correct responses based on the
PUT and GET requests sent when using "smt configure
--store=html://...". It is possible that the PUT request should be a
POST request. Currently, simulation records so only the configuration
works correctly (can't use "smt run" at this stage).

 * Add serialize method to ProjectModel for json output.

 * Add class level access to the database in UserModel so that db
   module calls are all done in the model classes.

 * Add script to delete all the data from the database for prototyping purposes.