Begin switch to MongoDB
Address #10

Switch from SQLAlchemy to MongoDB due to MongoDB's flexibility with
JSON. Decided to use Flask-MongoEngine and MongoEngine since this
seems quite mature and exposes a lot of the query stuff.
Flask-PyMongo was investigated, but seemed to have some issues.

Currently the API is switched off (commented out) until we have the
Project and Record models updated to use Mongo.

Pulled in a custom converter for objectids as they might be used a lot
in urls.