Use Sumatra client authentication for user
Fix #29

Create simple functions to authenticate the user using the Sumatra client's
existing infrastructure for authentication. This is the simplest short
term solution to at least send the user information from the client to
the server. This enables future work on the front end web view to go
ahead since the database can now be more easily populated with users,
projects and records.

This is a short term solution, in the long run a much better
authentication is required. This will require pervasive changes to the
client.

 * Create "common/tools" module for shared miscellaneous resources

 * Introduced "to_smt_json" methods to return the json format expected
   but the Sumatra client.

 * Removed "nickname" field from UserModel. The field is redundant and
   it seems better just to use the email.

 * "get_anonymous" is no longer needed as users are now authenticated.

 * Both API url endpoints are now authenticated