Display user and project pages using DataTables
Address #18

Front end display uses DataTables to display the list of a user's
projects and a list of the project's records. This is a basic
requirement to view the data in the database. A "get_datatable" method
was created on UserModel and ProjectModel. It returns lists that are
turned into json by the templates when passed to the Javascript.
Currently the project and user pages are using the same
templates. This needs to be modified slightly to distinguish them.