Improve DataTables and Psdash integration
Address #38

CSS updated in psdash.css to improve the look and feel of the
tables. Probably need to create a sumatra-cloud.css that overwrite the
psdash.css rather than changing it. Tables are auto sorted by the data
of last update

Removed "created_at" fields as they are automatically recorded by
mongo on creation of the model instance. The project's last_updated
field is created from the records last_updated fields. RecordModel
overwrites the save method to record when the record was last updated.